# Cdef inner funcs