--- a/setup.py
+++ b/setup.py
@@ -1,13 +1,9 @@
 """
 Fast intensity inference
 """
-from codecs import open
 from os.path import abspath, dirname, join
 
 from setuptools import setup, Extension
-
-# Build time requires numpy and Cython
-# Runtime just requires numpy and scipy
 import numpy
 from Cython.Build import cythonize
 
@@ -15,8 +11,13 @@ readme_path = join(abspath(dirname(__file__)), 'README.md')
 with open(readme_path, encoding='utf-8') as file:
     DESCRIPTION = file.read()
 
-extensions = [Extension('fast_intensity', ['fast_intensity.pyx'],
-              include_dirs=[numpy.get_include()])]
+extensions = [
+    Extension(
+        'fast_intensity',
+        ['fast_intensity.pyx'],
+        include_dirs=[numpy.get_include()],
+    )
+]
 
 # Run the build
 setup(
@@ -37,7 +38,6 @@ setup(
     maintainer_email='john.m.still@vumc.org',
     long_description=DESCRIPTION,
     long_description_content_type='text/markdown',
-
     # Install data
     zip_safe=False,
     include_package_data=True,