Before this update:
3.65 ms ± 57.3 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
23 ms ± 2.3 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)

After this update:
2.1 ms ± 49.8 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
13.1 ms ± 161 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

Timings produced on an OS X i9 6-core using `bench.py`.